Minor fixes.
* imap4d/imap4d.c: Avoid gcc warnings.
* libmailutils/tests/imapio.c (main): Fix format specifier.
* maidag/mailquota.c (dbm_retrieve_quota): Likewise.
* mail/quit.c (mail_mbox_close): Likewise.
* mail/mail.c (options): Fix improper initialization.
* mu/mu.h (mu_getans): Add proto.
* mu/util.c: Include netinet/in.h

+++ libmailutils/tests/imapio.c
@@ -82,7 +82,7 @@ main (int argc, char **argv)
       if (wc == 0)
 	break;
 
-      mu_stream_printf (mu_strerr, "%d\n", wc);
+      mu_stream_printf (mu_strerr, "%lu\n", (unsigned long) wc);
       for (i = 0; i < wc; i++)
 	{
 	  mu_stream_printf (mu_strerr, "%d: '%s'\n", i, wv[i]);
